Add files via upload
The Kubernetes and Cloud Native Associate (KCNA) certification exam is a must for each and everyone who wants to make foray into the Containers space. It’s a foundation level certification exam where the focus is on Kubernetes and cloud native framework. The competencies which would be covered as part of this course is as follows:
•	Cloud Native Architecture – 16%
•	Container Orchestration – 22%
•	Kubernetes Fundamentals – 46%
•	Cloud Native Application Delivery – 8%
•	Cloud Native Observability – 8%

Additional Information: https://medium.com/@prachetmohanty/kubernetes-and-cloud-native-associate-kcna-exam-in-a-nutshell-de18e5fdb6d9

Hope these notes may prove helpful in your certification preparation. All the best!!